requirements.txt
asgiref==3.2.5
boto3==1.12.43
botocore==1.15.43
certifi==2020.4.5.1
chardet==3.0.4
client==0.0.1
defusedxml==0.6.0
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.0.4
django-crispy-forms==1.9.0
django-filter==2.2.0
django-heroku==0.3.1
django-import-export==2.3.0
django-mysql==3.3.0
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
dnspython==1.16.0
docutils==0.15.2
et-xmlfile==1.0.1
gunicorn==20.0.4
idna==2.9
jdcal==1.4.1
jmespath==0.9.5
joblib==0.17.0
Markdown==3.2.1
MarkupPy==1.14
mysql==0.0.2
mysql-connector==2.2.9
mysql-connector-python==8.0.19
mysqlclient==1.4.6
numpy==1.19.2
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.3
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.5
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2019.3
PyYAML==5.3.1
requests==2.23.0
s3transfer==0.3.3
scikit-learn==0.23.2
scipy==1.5.3
six==1.14.0
sklearn==0.0
sqlparse==0.3.1
tablib==2.0.0
threadpoolctl==2.1.0
urllib3==1.25.9
whitenoise==5.2.0
xlrd==1.2.0
xlwt==1.3.0
